Finshed preprocessing and made notes on how to do data streaming
This is unprofesional, but todays work is not great. I was uplate
trying to debug why I could not read movies on aws... it was such
a silly reason. But, I know how I am going to make the data streamer,
I just have to do it and then I can make a KFold to go on top of it.
Maybe I will get to it after I take a nap.